backup-system/requirements.txt
# Core dependencies
boto3>=1.26.0
psycopg2-binary>=2.9.0
PyYAML>=6.0
click>=8.1.0
schedule>=1.2.0
python-json-logger>=2.0.7
requests>=2.28.0
cryptography>=3.4.0
hashlib-compat>=1.0.0

# Development dependencies
pytest>=7.0.0
pytest-cov>=4.0.0
black>=22.0.0
flake8>=5.0.0
mypy>=1.0.0

# Optional dependencies
watchdog>=2.1.0  # Para monitoramento de arquivos
tqdm>=4.64.0  # Para barras de progresso
colorama>=0.4.0  # Para output colorido